+++ govcd/vm_test.go
@@ -580,14 +580,14 @@ func (vcd *TestVCD) Test_AddMetadataOnVm(check *C) {
 	vm.VM = &vmType
 
 	// Add metadata
-	task, err := vapp.AddMetadata("key", "value")
+	task, err := vm.AddMetadata("key", "value")
 	check.Assert(err, IsNil)
 	err = task.WaitTaskCompletion()
 	check.Assert(err, IsNil)
 	check.Assert(task.Task.Status, Equals, "success")
 
 	// Check if metadata was added correctly
-	metadata, err := vcd.vapp.GetMetadata(vm.VM.HREF)
+	metadata, err := vm.GetMetadata()
 	check.Assert(err, IsNil)
 	check.Assert(metadata.MetadataEntry[0].Key, Equals, "key")
 	check.Assert(metadata.MetadataEntry[0].TypedValue.Value, Equals, "value")
@@ -611,7 +611,7 @@ func (vcd *TestVCD) Test_DeleteMetadataOnVm(check *C) {
 	vm.VM = &vmType
 
 	// Remove metadata
-	task, err := vapp.DeleteMetadata("key")
+	task, err := vm.DeleteMetadata("key")
 	check.Assert(err, IsNil)
 	err = task.WaitTaskCompletion()
 	check.Assert(err, IsNil)